requirements.txt
Django==1.6
pymongo
mysql-python
django-registration==1.0